requirements.txt
# --------- pytorch --------- #
torch>=2.0.0
torchvision>=0.15.0
lightning[extra]>=2.0.0
torchmetrics>=0.11.4

# --------- model zoo -------#
transformers

# ------- tensor utils ------#
einops

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- deployment ------ #
fastapi

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
black
# sh            # for running bash commands in some tests (linux/macos only)

dvc==3.1.0